// example3/single-spa-vite/vendor-account-ui/src/middleware/invoice/invoice.types.ts
export const InvoiceOperations = {
  MERCH_ACCEPT: 'merchAccept',
  MERCH_REJECT: 'merchReject',
  MERCH_SHIPMENT: 'merchShipment',
  MERCH_SPLIT: 'merchSplit',
  MERCH_1C: 'merch1C',
  PRINT_ACT: 'printAct',
  DOWNLOAD_UPD: 'downloadUPD',
  UPDATE_CDK: 'updateCDK',
  LABEL_PRINT: 'labelPrint',
  MERCH_KDK: 'merchKdk',
  MERCH_PRINT_LABEL: 'merchPrintlabel',
  MERCH_SHIP_KDK: 'merchShipkdk',
} as const;
export const InvoiceProcedure = {
  [InvoiceOperations.MERCH_ACCEPT]: '40105512',
  [InvoiceOperations.MERCH_REJECT]: '40105511',
  [InvoiceOperations.MERCH_SHIPMENT]: '40105510',
  [InvoiceOperations.MERCH_SPLIT]: '40109121',
  [InvoiceOperations.MERCH_1C]: '40108035',
  [InvoiceOperations.PRINT_ACT]: '40048481',
  [InvoiceOperations.DOWNLOAD_UPD]: '40048118',
  [InvoiceOperations.UPDATE_CDK]: '40109047',
  [InvoiceOperations.LABEL_PRINT]: '20048003',
  [InvoiceOperations.MERCH_KDK]: '',
  [InvoiceOperations.MERCH_PRINT_LABEL]: '',
  [InvoiceOperations.MERCH_SHIP_KDK]: '40105614',
} as const;
// ...
export interface IReturnAdapterInvoiceBody {
  lineNum: number;
  rowId: string;
  id: string;
  productCode: string;
  clientProductCode: string;
  productName: string;
  manufacturer: string;
  article: string;
  productCount: number;
  price: string;
  sum: string;
  deliveryTime: string;
  requiredDeliveryDate: string;
  supplierDate: string;
  rejectionReason: string;
  comment: string;
  measure: string;
}
// ...
interface IInvoiceProcedureAccept {
  supplierDate: OrderRowListsType['supplierDate'];
}
interface IInvoiceProcedureReject extends Partial<IInvoiceProcedureAccept> {
  rejectionReason: OrderRowListsType['rejectionReason'];
  comment: OrderRowListsType['comment'];
}

export interface IInvoiceBodyParams {
  [InvoiceProcedure.merchAccept]: Record<string, IInvoiceProcedureAccept>;
  [InvoiceProcedure.merchReject]: Record<string, IInvoiceProcedureReject>;
  [InvoiceProcedure.merchSplit]: Record<string, boolean>;
  [InvoiceProcedure.updateCDK]: Record<string, number | string>[];
}